requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
pandas==2.1.3
faiss-cpu==1.7.4
numpy==1.26.2
python-multipart==0.0.6
pydantic==2.5.2
gspread==6.2.0
oauth2client==4.1.3
jinja2==3.1.2
flask>=2.2.2
flask-cors>=3.0.10
requests>=2.28.1
tiktoken>=0.4.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.90.0
